Expose queue mocks for testing outside of go-base
It's needed to mock the queue in services that use this package.
